Version 1.5.4
git-svn-id: https://plugins.svn.wordpress.org/wordpress-backup-to-dropbox/trunk@718379 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ Classes/class-wp-backup-database.php
@@ -2,7 +2,7 @@
 /**
  * A class with functions the perform a backup of WordPress
  *
- * @copyright Copyright (C) 2011-2012 Michael De Wildt. All rights reserved.
+ * @copyright Copyright (C) 2011-2013 Michael De Wildt. All rights reserved.
  * @author Michael De Wildt (http://www.mikeyd.com.au/)
  * @license This program is free software; you can redistribute it and/or modify
  *          it under the terms of the GNU General Public License as published by
@@ -39,8 +39,11 @@ abstract class WP_Backup_Database {
 	}
 
 	public function remove_file() {
-		foreach (glob($this->get_file(false) . '*') as $file)
-			unlink($file);
+		$files = glob($this->get_file(false) . '*');
+		if ($files) {
+			foreach ($files as $file)
+				unlink($file);
+		}
 	}
 
 	private function set_wait_timeout() {
